% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss_tree.R
\name{gamlss_tree}
\alias{gamlss_tree}
\title{Gamlss models tree}
\usage{
gamlss_tree(form, datos, n_dist_mod = 4, var_sel = "aicmodelo",
  steps = 2, porc_entre = 0.8, committess = 1,
  nom_dist = c("exGAUS", "GIG", "GG", "BCCGo", "BCPEo", "GA", "GB2",
  "BCTo", "WEI3", "LOGNO", "EXP", "PARETO2", "IG", "IGAMMA", "NO"),
  cyc = 50, prueba_hip = TRUE, acepta_h = FALSE, type = "counts",
  arbol_activo = TRUE)
}
\arguments{
\item{form}{a formula object, with the response on the left of an ~ operator, 
and the terms, separated by + operators, on the right.}

\item{datos}{a data frame.}

\item{n_dist_mod}{distributions number}

\item{var_sel}{selection parameter.}

\item{steps}{steps in stepwise.}

\item{porc_entre}{percentage in training sample .}

\item{committess}{boosting iterations}

\item{nom_dist}{gamlss distributions.}

\item{cyc}{gamlss iterations.}

\item{prueba_hip}{goodness of fit for user distributions.}

\item{acepta_h}{goodness of fit for response variable.}

\item{type}{type of distribution.}

\item{arbol_activo}{decision tree to start.}
}
\value{
\code{gamlss_tree} gives nine results: arboles, modelosxnodo, 
valor_nodo, dis_sel, nodos_train, Form, datos_test, nodo_prueba, and 
nodos_ajuste_cov.
}
\description{
Gamlss models tree.
}
\details{
The gamlss models tree use decision tree and gamlss models.
}
\examples{
require(COUNT)
data(azpro)
model_amg<-gamlss_tree(los~.,datos=azpro)
model_amg$arboles
 
}
