% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_gamlss_tree.R
\name{pred_gamlss_tree}
\alias{pred_gamlss_tree}
\title{Prediction gamlss tree}
\usage{
pred_gamlss_tree(newdata = newdata, objeto)
}
\arguments{
\item{newdata}{a data.frame.}

\item{objeto}{gamlss_tree list.}
}
\value{
\code{pred_gamlss_tree} gives a data.frame with data, pred_arbol and pred_gam.
}
\description{
Prediction gamlss tree.
}
\examples{
require(COUNT)
data(azpro)
model_amg<-gamlss_tree(los~.,datos=azpro)
pred<-pred_gamlss_tree(objeto=model_gam,newdata=azpro)

}
