% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejemplo.R
\name{sm}
\alias{sm}
\title{Separar muestra}
\usage{
sm(datos, perc)
}
\arguments{
\item{datos}{conjunto de datos.}

\item{perc}{percentil.}
}
\value{
retorna la muestra divididaa
}
\description{
El objetivo es separa las muestra.
}
\details{
La función divide la muestra en dos partes.
}
\examples{
## TDivisión de la función uno
x <- cbind(rnorm(100, mean=20, sd=1), rnorm(100, mean=10, sd=0.5))
sm(x, 0.90)
}
