% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_sample.R
\name{split_sample}
\alias{split_sample}
\title{Split sample}
\usage{
split_sample(datos, perc)
}
\arguments{
\item{datos}{a data.frame}

\item{perc}{percentile of training sample.}
}
\value{
\code{split_sample} gives two samples
}
\description{
Split sample in test and train.
}
\examples{
require(COUNT)
data(azpro)
sample <- split_sample(datos=azpro, perc=0.8)

}
