% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_dist.R
\name{test_dist}
\alias{test_dist}
\title{Test distributions}
\usage{
test_dist(y, type = "realplus", extra = "NO", signif = 0.05)
}
\arguments{
\item{y}{a formula object, with the response on the left of an ~ operator, 
and the terms, separated by + operators, on the right.}

\item{type}{type of distribution.}

\item{extra}{additional distributions.}

\item{signif}{significance level.}
}
\value{
\code{test_dist} gives a data.frame with p-value, AIC, SBC and result of 
goodness of fitness test.
}
\description{
Test distributions.
}
\examples{
require(COUNT)
data(azpro)
test<-test_dist(y=azpro$los)
test

}
