% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss_tree.R
\name{gamlss_tree}
\alias{gamlss_tree}
\title{The gamlss tree methodology.}
\usage{
gamlss_tree(form, datos, n_dist_mod = 4, var_sel = "aicmodelo",
  steps = 2, porc_entre = 0.8, committess = 1,
  nom_dist = c("exGAUS", "GIG", "GG", "BCCGo", "BCPEo", "GA", "GB2",
  "BCTo", "WEI3", "LOGNO", "EXP", "PARETO2", "IG", "IGAMMA", "NO"),
  cyc = 50, prueba_hip = TRUE, acepta_h = FALSE, type = "counts",
  arbol_activo = TRUE)
}
\arguments{
\item{form}{percentiles 1.}

\item{n_dist_mod}{percentiles3.}

\item{var_sel}{percentiles4.}

\item{steps}{percentiles5.}

\item{porc_entre}{percentiles6.}

\item{committess}{percentiles7.}

\item{nom_dist}{percentile8.}

\item{cyc}{percentiles9.}

\item{prueba_hip}{percentiles10.}

\item{acepta_h}{percentiles11.}

\item{type}{percentiles12.}

\item{arbol_activo}{percentiles13.}

\item{perc}{percentiles 2.}
}
\value{
\code{gamlss_tree} gives two columns.
}
\description{
The function gamlss tree.
}
\details{
The gamlss_tree generates two dataset.
}
\examples{
## The gamlss_tree function
require(COUNT)
data(azpro)
model_amg<-gamlss_tree(los~.,datos=azpro)
model_amg$arboles


}
